Housing Association App.
Is a web application for managing housing associations. Project was built under Spring Model View Controller principles. For backend processes is responsible Java Object Orietned Programming Language. User interface was built with help of HTML, CSS, JQuery and Bootstrap library. With help of Thymeleaf data is extracted and shown on a website. MySQL database was build under Spring Data, Hibernate technologies. The main requirements of the app are following points:
1. Adding and displaying list of Housing Associations.
2. Adding and displaying list of all flats in all Housing Associations.
3. Displaying details of each Housing Association: sum of areas of all flats, list of all occupants, list of all flats of particular Housing Association.
4. Adding occupant to flat.
5. Displaying details of flat with list of occupants.
6. Deleting and editing data of Housing Association.
7. Deleting and editing data of flat.
8. Deleting and editing data of occupant.